+++ zktools/tests/__init__.py
@@ -1,1 +1,15 @@
-#
+import unittest
+
+__all__ = ['TestBase']
+
+
+connection = []
+
+
+class TestBase(unittest.TestCase):
+    @property
+    def conn(self):
+        from zktools.connection import ZkConnection
+        if not connection:
+            connection.append(ZkConnection())
+        return connection[0]