+++ CHANGES.rst
@@ -15,6 +15,12 @@ Features
 
 - Configuration object to load/store configuration data in Zookeeper.
 
+Backward Incompatibilities
+**************************
+
+- SharedZkLock has been refactored into ZkWriteLock and ZkReadLock.
+- ``reconnect`` is no longer an option to ZkConnection.
+
 
 0.1 (11/22/2011)
 ----------------